requirements.txt
PILLOW
face_recognition
git+https://github.com/ageitgey/face_recognition_models
setuptools